requirements.txt
easysnmp==0.2.5
pyTelegramBotAPI==4.0.0
influxdb==5.3.1
python-dotenv==0.19.0
rich==10.9.0